requirements.txt
PyMuPDF
pillow
pytesseract
